server/requirements.txt
pycryptodome==3.14.1
               
